Modified the locations fields to integer

+++ database/migrations/2018_02_11_064917_create_hostels_table.php
@@ -18,8 +18,8 @@ class CreateHostelsTable extends Migration
             $table->string('name');
             $table->text('description');
             $table->integer('totalRooms')->default(0);
-            $table->integer('latitude');
-            $table->integer('longitude');
+            $table->double('latitude');
+            $table->double('longitude');
             $table->double('price')->default(0);
             $table->string('contact');
             $table->double('deposit')->default(0);
